refactor(admin): split buttondown card modules
separate the content-detail Buttondown card into typed state, action orchestration, and presentational components so the entry component stays focused on rendering the current mode.

verify: pnpm exec prettier --check apps/admin/src/app/content/[id]/buttondown-card.tsx apps/admin/src/app/content/[id]/buttondown-card-state.ts apps/admin/src/app/content/[id]/buttondown-card-types.ts apps/admin/src/app/content/[id]/buttondown-card-ui.tsx

verify: git diff --check -- apps/admin/src/app/content/[id]/buttondown-card.tsx apps/admin/src/app/content/[id]/buttondown-card-state.ts apps/admin/src/app/content/[id]/buttondown-card-types.ts apps/admin/src/app/content/[id]/buttondown-card-ui.tsx

verify: pnpm --filter @anipotts/admin typecheck

verify: pnpm --filter @anipotts/admin lint

verify: pnpm --filter @anipotts/admin build

verify: pnpm validate
