+++ apps/admin/src/app/content/[id]/buttondown-card-types.ts
@@ -1,0 +1,11 @@
+export interface ButtondownState {
+  emailId?: string;
+  status?: string;
+  subject?: string;
+  body?: string;
+}
+
+export type ButtondownFeedback = {
+  type: "success" | "error";
+  message: string;
+} | null;
